split main.py functionalities into separate files
