+++ .gitignore
@@ -2,4 +2,4 @@ env/
 attachments/
 .idea/
 venv/
-__pycache__/
+src/__pycache__/
