+++ setup.cfg
@@ -10,7 +10,7 @@ ignore=W503,W504,E203,E731,E501
 
 [isort]
 line_length = 88
-sections=FUTURE,STDLIB,THIRDPARTY,LOCALFOLDER
+sections=FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
 default_section=THIRDPARTY
 multi_line_output=3
 include_trailing_comma=true
